chore: rm chance of hight ce costs

+++ justfile
@@ -271,21 +271,32 @@ backend-build:
         echo "✅ Done: backend/$app_name.zip"
     done
 
-
-start:
+seed:
     #!/usr/bin/env bash
-    # if ! docker info > /dev/null 2>&1; then
-    #   echo "Error: Docker daemon is not running. Please start Docker and try again." >&2
-    #   exit 1
-    # fi
+    set -e
+
+    RED='\033[0;31m'
+    NC='\033[0m' # No Color
+
+    echo -e "${RED}⚠️  WARNING: This operation will call the AWS Cost Explorer API and may incur charges beyond the free tier.${NC}"
+    read -p "Do you want to continue? (y/n): " confirm
+    if [[ "$confirm" != "y" ]]; then
+      echo "❌ Aborted."
+      exit 1
+    fi
 
     python3 -m venv venv
     source venv/bin/activate
     pip install python-dotenv boto3
+    export ENVIRONMENT_NAME=prod
+    export PROJECT_NAME=chrispsheehan-webstack 
+    export PUBLIC_DIR=${PWD}/frontend/public 
+    python backend/local_runner.py
 
-    ENVIRONMENT_NAME=prod PROJECT_NAME=chrispsheehan-webstack PUBLIC_DIR=${PWD}/frontend/public python backend/local_runner.py
 
-    # npm i --prefix frontend
-    # docker compose up -d
-    # docker compose logs -f &
-    # npm run dev --prefix frontend
+start:
+    #!/usr/bin/env bash
+    npm i --prefix frontend
+    docker compose up -d
+    docker compose logs -f &
+    npm run dev --prefix frontend
